--- internal/initwd/inode_windows.go
@@ -1,8 +1,0 @@
-// +build windows
-
-package initwd
-
-// no syscall.Stat_t on windows, return 0 for inodes
-func inode(path string) (uint64, error) {
-	return 0, nil
-}
